+++ Havtorn/ProjectSetup/AdditionalConfigStart.py
@@ -1,11 +1,7 @@
 import os
 import subprocess
-import platform
 
-from ConfigureMSVC import MSVCConfiguration as MSVCRequirements
-MSVCRequirements.Validate()
-
-from ConfigurePython import PythonConfiguration as PythonRequirements
+from AdditionalConfigPython import PythonConfiguration as PythonRequirements
 PythonRequirements.Validate()
 
 subprocess.call([os.path.abspath("./GenerateProjectFiles.bat"), "nopause"])
