+++ Havtorn/SetupProject.bat
@@ -24,10 +24,10 @@ py --version>nul 2>&1
 set pyCmdError=%errorlevel%
 if %pythonCmdError% NEQ 0 (
     if %pyCmdError% NEQ 0 goto :ON_REQUIREMENT_ERROR
-        py ConfigureStartPoint.py
+        py AdditionalConfigStart.py
         goto :COMPLETE
 )
-python ConfigureStartPoint.py
+python AdditionalConfigStart.py
 
 :COMPLETE
 echo.
